* subversion/tests/libsvn_subr/named_atomic-test.c   (calibrate_iterations): silence a precision loss warning
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1326617 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/tests/libsvn_subr/named_atomic-test.c
@@ -267,7 +267,7 @@ calibrate_iterations(apr_pool_t *pool, int count)
       start = apr_time_now();
       SVN_ERR(run_procs(pool, TEST_PROC, count, calib_iterations));
 
-      taken = apr_time_now() - start;
+      taken = (double)(apr_time_now() - start);
     }
 
   /* scale that to 1s */
